move customer request to correct namespace add category behat tests
