+++ src/Request/Customers/CustomerEmailTokenRequest.php
@@ -1,10 +1,10 @@
 <?php
 /**
  * @author @ct-jensschulze <jens.schulze@commercetools.de>
- * @created: 11.02.15, 16:42
+ * @created: 12.02.15, 12:34
  */
 
-namespace Sphere\Core\Request\Customer;
+namespace Sphere\Core\Request\Customers;
 
 use Sphere\Core\Client\HttpMethod;
 use Sphere\Core\Client\JsonRequest;
@@ -12,32 +12,30 @@ use Sphere\Core\Request\AbstractUpdateRequest;
 use Sphere\Core\Response\SingleResourceResponse;
 
 /**
- * Class CustomerPasswordChangeRequest
- * @package Sphere\Core\Request\Customer
- * @method static CustomerPasswordChangeRequest of($id, $version, $currentPassword, $newPassword)
+ * Class CustomerEmailTokenRequest
+ * @package Sphere\Core\Request\Customers
+ * @method static CustomerEmailTokenRequest of(string $id, int $version, int $ttlMinutes)
  */
-class CustomerPasswordChangeRequest extends AbstractUpdateRequest
+class CustomerEmailTokenRequest extends AbstractUpdateRequest
 {
     const ID = 'id';
-    const CURRENT_PASSWORD = 'currentPassword';
-    const NEW_PASSWORD = 'newPassword';
+    const TTL_MINUTES = 'ttlMinutes';
 
-    protected $currentPassword;
-    protected $newPassword;
-
+    /**
+     * @var int
+     */
+    protected $ttlMinutes;
     /**
      * @param string $id
      * @param string $version
-     * @param string $currentPassword
-     * @param string $newPassword
+     * @param int $ttlMinutes
      */
-    public function __construct($id, $version, $currentPassword, $newPassword)
+    public function __construct($id, $version, $ttlMinutes)
     {
         parent::__construct(CustomersEndpoint::endpoint(), $id, $version);
         $this->setId($id);
         $this->setVersion($version);
-        $this->currentPassword = $currentPassword;
-        $this->newPassword = $newPassword;
+        $this->ttlMinutes = $ttlMinutes;
     }
 
     /**
@@ -46,7 +44,7 @@ class CustomerPasswordChangeRequest extends AbstractUpdateRequest
      */
     protected function getPath()
     {
-        return (string)$this->getEndpoint() . '/password';
+        return (string)$this->getEndpoint() . '/email-token';
     }
 
     /**
@@ -58,8 +56,7 @@ class CustomerPasswordChangeRequest extends AbstractUpdateRequest
         $payload = [
             static::ID => $this->getId(),
             static::VERSION => $this->getVersion(),
-            static::CURRENT_PASSWORD => $this->currentPassword,
-            static::NEW_PASSWORD => $this->newPassword
+            static::TTL_MINUTES => $this->ttlMinutes,
         ];
         return new JsonRequest(HttpMethod::POST, $this->getPath(), $payload);
     }
