+++ src/Request/Customers/CustomerFetchByIdRequest.php
@@ -4,13 +4,13 @@
  * @created: 11.02.15, 14:17
  */
 
-namespace Sphere\Core\Request\Customer;
+namespace Sphere\Core\Request\Customers;
 
 use Sphere\Core\Request\AbstractFetchByIdRequest;
 
 /**
  * Class CustomerFetchByIdRequest
- * @package Sphere\Core\Request\Customer
+ * @package Sphere\Core\Request\Customers
  * @method static CustomerFetchByIdRequest of(string $id)
  */
 class CustomerFetchByIdRequest extends AbstractFetchByIdRequest
