+++ src/Request/Customers/CustomersQueryRequest.php
@@ -4,14 +4,14 @@
  * @created: 11.02.15, 14:19
  */
 
-namespace Sphere\Core\Request\Customer;
+namespace Sphere\Core\Request\Customers;
 
 
 use Sphere\Core\Request\AbstractQueryRequest;
 
 /**
  * Class CustomersQueryRequest
- * @package Sphere\Core\Request\Customer
+ * @package Sphere\Core\Request\Customers
  * @method static CustomersQueryRequest of()
  */
 class CustomersQueryRequest extends AbstractQueryRequest
